Login Backend Completed
Login Backend Completed and up to date with the current database schema as of 12-11 and correctly logs a SUccessful or failed login. Temp fix applied to failed login because the database setup doesnt allow for a null S_ID. we need to figure out how we want to address this moving forward.

Login Page Decorated with light styling. purely for my own viewing pleasure on my old eyes.